Renames bricking_assessment module
